schema: add exception fields to schema
Problem: Flux contains details for the reason of a job failure (known
as an "exception") that could be useful for both monitoring and
visualization, but they are not listed in the Job Completion Log
Schema.

Add two new fields to the Job Completion Log Schema:

job.exception_type: The type of exception that was raised on a job.

job.exception_note: Any notes or additional details containing the
reason for a job exception.