schema: add "Notes" section for some fields
Problem: The HPC Job Schedular Event Schema has fields that are not
always guaranteed to be populated from the Flux job side, and it would
help to give more context to these fields.

Add a "Notes" section to the bottom of the schema table with some
clarifications on why certain event.* fields might not always be
populated.